CPG-15 created GET doctor profile endpoint
Implemented the GET doctor profile endpoint in doctor.js under the controller folder.
Retrieved the doctor from the database using doctorId and removed unwanted fields.
Sent the profile data as a response.
Modified the auth.js file to handle errors when selecting roles and if a role is not found.

Co-Authored-By: Venul Minsara <188411232+menulVIN@users.noreply.github.com>